2024_10_19

+++ OtpishAI_Backend_Codes/src/main/java/otpishAI/otpishAI_Backend/service/CartService.java
@@ -3,25 +3,27 @@ package otpishAI.otpishAI_Backend.service;
 import lombok.AllArgsConstructor;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
+import otpishAI.otpishAI_Backend.dto.CartDTO;
+import otpishAI.otpishAI_Backend.dto.CartResponseDTO;
 import otpishAI.otpishAI_Backend.entity.Cart;
 import otpishAI.otpishAI_Backend.entity.ProductDetail;
-import otpishAI.otpishAI_Backend.entity.Review;
 import otpishAI.otpishAI_Backend.repository.CartRepository;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
 public class CartService {
     private final CartRepository cartRepository;
 
-    public void updateCart(String username, ProductDetail productDetail){
+    public void addOrUpdateCart(String username, ProductDetail productDetail, String productCode){
         Optional<Cart> existCart = cartRepository.findByUsernameAndDetailCode(username, productDetail.getDetailCode());
 
         if(existCart.isPresent()){
             Cart updateCart = existCart.get();
-            updateCart.setQuantity(updateCart.getQuantity() + 1);
+            updateCart.setQuantity(updateCart.getQuantity() + 1L);
             cartRepository.save(updateCart);
         } else {
             Cart newCart= new Cart();
@@ -29,14 +31,38 @@ public class CartService {
             newCart.setDetailCode(productDetail.getDetailCode());
             newCart.setOPrice(productDetail.getOPrice());
             newCart.setRPrice(productDetail.getRPrice());
-            newCart.setQuantity((long)1);
+            newCart.setQuantity(1L);
+            newCart.setProductCode(productCode);
             cartRepository.save(newCart);
+        }
+    }
+
+    public Boolean minusCart(Long cartNum){
+        try{
+            Cart cart = cartRepository.findByCartNum(cartNum);
+            if(cart == null){
+                return false;
+            }
+            if(cart.getQuantity() <= 1) {
+                cartRepository.deleteByCartNum(cartNum);
+            }
+            else{
+                cart.setQuantity(cart.getQuantity() - 1L);
+                cartRepository.save(cart);
+            }
+            return true;
+        }catch (DataAccessException e){
+            return false;
         }
     }
 
 
     public Boolean deleteCart(Long cartNum){
         try{
+            Cart cart = cartRepository.findByCartNum(cartNum);
+            if(cart == null){
+                return false;
+            }
             cartRepository.deleteByCartNum(cartNum);
             return true;
         }catch (DataAccessException e){
@@ -44,9 +70,36 @@ public class CartService {
         }
     }
 
-    public List<Cart> cartssByUsername(String username){
 
-        return cartRepository.findAllByUsernameOrderByCartNum(username);
+    public Boolean plusCart(Long cartNum){
+       try{
+            Cart cart = cartRepository.findByCartNum(cartNum);
+            if(cart == null){
+                return false;
+            }
+            cart.setQuantity(cart.getQuantity() + 1L);
+            cartRepository.save(cart);
+            return true;
+        }catch (DataAccessException e){
+            return false;
+        }
+    }
+
+    public CartResponseDTO cartsByUsername(String username){
+
+        List<Cart> cartList = cartRepository.findAllByUsernameOrderByCartNum(username);
+        List<CartDTO> cartDTOS = cartList.stream()
+                .map(cart -> new CartDTO(cart))
+                .collect(Collectors.toList());
+        CartResponseDTO cartResponseDTO = new CartResponseDTO();
+        cartResponseDTO.setCartDTOS(cartDTOS);
+        cartResponseDTO.setCartCNT(cartDTOS.size());
+
+        return cartResponseDTO;
+    }
+
+    public Cart cartFind(String detailCode){
+        return cartRepository.findByDetailCode(detailCode);
     }
